New game in-game rework
Fixed some stuff that made generating new game using the * button in-game a little weird. The window would just freeze leaving the user wandering what in the world is going on. Now when the user presses the button a neat overlay pops up informing the user that generation is on underway. A timer is there to show the user the app isn't frozen. And if you hold shift while pressing the button the generation is animated. Which is...neat, I-I guess?

+++ src/main/java/cz/krejcar25/sudoku/ui/Clock.java
@@ -13,7 +13,7 @@ public class Clock extends Drawable {
     private Digit secD;
     private Digit secU;
 
-    public Clock(Applet applet, int x, int y, String name) {
+    public Clock(Applet applet, float x, float y, String name) {
         super(applet, x, y, baseWidth, baseHeight);
         this.timer = new Timer(name);
         this.minD = new Digit(getApplet(), 0, 0);
@@ -22,7 +22,7 @@ public class Clock extends Drawable {
         this.secU = new Digit(getApplet(), minD.width + minU.width + 50 + secD.width, 0);
     }
 
-    public Clock(Applet applet, int x, int y, Timer timer) {
+    public Clock(Applet applet, float x, float y, Timer timer) {
         super(applet, x, y, baseWidth, baseHeight);
         this.timer = timer;
         this.minD = new Digit(getApplet(), 0, 0);
